Adding background file linking.

+++ ush/gdas/soca/__init__.py
@@ -355,8 +355,8 @@ class SOCA:
 
         # Define the background forecast times relative to the
         # respective analysis time.
-        ntimes = float(bkgrds_config_obj.analysis_interval_seconds /
-                       bkgrds_config_obj.bkgrd_interval_seconds) + 1.0
+        ntimes = int((bkgrds_config_obj.analysis_interval_seconds /
+                      bkgrds_config_obj.bkgrd_interval_seconds) + 1.0)
 
         offset_seconds_list = [offset for offset in numpy.linspace(
             ((bkgrds_config_obj.analysis_interval_seconds/2.0) -
